rename package to "abstrakt"
